fix multiclient e2e
